+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Reverted
+
+- Remove `listHistoricalBalances` wallet method to avoid confusion.
+
 ## [0.1.0] - 2024-08-22
 
 ### Added